Plan: plans/PR-FAQ-Deflection-Report-Term-File.md
Slice phase: Product polish

Move the documentation-term-file parser into the shared FAQ CLI rules module and wire the customer-facing deflection report CLI to the same parser.

Intentional:

- No new rule-file schema for documentation terms; rule files stay limited to intent and vocabulary-gap rules.

- No report Markdown or hosted execute route changes; this is offline report CLI ergonomics.

Deferred:

- Hosted/deployed deflection report validation remains deferred until the operator-provisioned Atlas host and JWT are available.

Parked hardening:

- None.

Verification:

- python -m py_compile scripts/content_ops_faq_cli_rules.py scripts/build_extracted_ticket_faq_markdown.py scripts/build_content_ops_deflection_report.py tests/test_content_ops_deflection_report.py tests/test_extracted_ticket_faq_markdown.py

- python -m pytest tests/test_content_ops_deflection_report.py tests/test_extracted_ticket_faq_markdown.py -q -k 'deflection_report_cli or documentation_term_file'

- python -m pytest tests/test_content_ops_deflection_report.py -q

- python -m pytest tests/test_extracted_ticket_faq_markdown.py -q

- python scripts/audit_plan_doc.py plans/PR-FAQ-Deflection-Report-Term-File.md

- python scripts/audit_plan_code_consistency.py plans/PR-FAQ-Deflection-Report-Term-File.md

- git diff --check
